# backend/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spice 'N' Rice API",
    "version": "1.0.0",
    "description": "All prices are integer USD cents. Mutations use same-origin HttpOnly cookies. Staff can edit operational records; only owners can delete records and change settings. Responses use Cache-Control: no-store."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8087/api/v1"
    },
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/menu": {
      "get": {
        "summary": "Visible categories and available menu items",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "name": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/MenuItem"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/inquiries": {
      "post": {
        "summary": "Submit contact or catering inquiry; five requests per minute per client",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inquiry"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Issue 15-minute access JWT and seven-day refresh cookies",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "role": {
                      "enum": [
                        "owner",
                        "staff"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Rotate refresh token and access JWT",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "name": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke session and clear cookies",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "name": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Current user",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "role": {
                      "enum": [
                        "owner",
                        "staff"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/admin/{resource}": {
      "parameters": [
        {
          "name": "resource",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "items",
              "categories",
              "catering",
              "settings",
              "testimonials",
              "inquiries",
              "media"
            ]
          }
        }
      ],
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "List records in display order",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "name": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      },
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Create record (media uses upload; settings restricted to owner)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/MenuItem"
                  },
                  {
                    "$ref": "#/components/schemas/Category"
                  },
                  {
                    "$ref": "#/components/schemas/CateringItem"
                  },
                  {
                    "$ref": "#/components/schemas/Settings"
                  },
                  {
                    "$ref": "#/components/schemas/Testimonial"
                  },
                  {
                    "$ref": "#/components/schemas/Inquiry"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/admin/bulk-prices": {
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Transactional bulk update; 100 basis points = 1%. Category 0 selects all.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_id": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "basis_points": {
                    "type": "integer",
                    "minimum": -10000,
                    "maximum": 100000
                  },
                  "set_cents": {
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 10000000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "confirm": {
                    "const": true
                  }
                },
                "required": [
                  "confirm"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Menu, unread inquiry, and weekly price change counts",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "name": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "MenuItem": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "category_id": {
            "type": "integer",
            "minimum": 1
          },
          "price_cents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000000
          },
          "price": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "veg": {
            "type": "boolean"
          },
          "spicy": {
            "type": "boolean"
          },
          "featured": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "price_cents",
          "category_id"
        ]
      },
...
    }
  }
}